Move RFF adapters to EspressoForm
