+++ assets/src/application/ui/forms/espressoForm/adapters/DateTimePicker/DatePicker.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { parse } from 'date-fns';
 import DatePickerBase from './DatePickerBase';
-import { CONVERT_TO_MOMENT_DATE_FORMAT } from '../../../constants';
+import { CONVERT_TO_MOMENT_DATE_FORMAT } from '@appConstants/dateFnsFormats';
 
 const DatePicker = ({ input, ...rest }) => {
 	return (
