+++ assets/src/application/ui/forms/espressoForm/adapters/DateTimePicker/TimePicker.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { parse } from 'date-fns';
 import TimePickerBase from './TimePickerBase';
-import { CONVERT_TO_MOMENT_TIME_FORMAT } from '../../../constants';
+import { CONVERT_TO_MOMENT_TIME_FORMAT } from '@appConstants/dateFnsFormats';
 
 const TimePicker = ({ input, ...rest }) => {
 	return (
