+++ assets/src/application/ui/forms/espressoForm/adapters/Switch.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Switch } from '@infraUI/inputs';
-import { FieldRendererProps } from '../forms/espressoForm';
+import { FieldRendererProps } from '../types';
 
 const SwitchField: React.FC<FieldRendererProps> = ({ input, meta: { error, submitError }, ...rest }) => {
 	return <Switch {...input} isChecked={Boolean(input.value)} isInvalid={error || submitError} {...rest} />;
