+++ assets/src/application/ui/forms/espressoForm/adapters/Text.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { TextInput } from '@infraUI/inputs';
-import { FieldRendererProps } from '../forms/espressoForm';
+import { FieldRendererProps } from '../types';
 
 const Text: React.FC<FieldRendererProps> = ({ htmlType = 'text', input, meta: { error, submitError }, ...rest }) => {
 	return <TextInput {...input} isInvalid={error || submitError} {...rest} />;
