+++ assets/src/application/ui/forms/espressoForm/adapters/TextArea.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { TextArea as AdaptedTextArea } from '@infraUI/inputs';
-import { FieldRendererProps } from '../forms/espressoForm';
+import { FieldRendererProps } from '../types';
 
 const TextArea: React.FC<FieldRendererProps> = ({ input, meta: { error, submitError }, ...rest }) => {
 	return <AdaptedTextArea {...input} isInvalid={error || submitError} {...rest} />;
