+++ assets/src/application/ui/input/index.ts
@@ -1,14 +1,4 @@
 export { default as CurrencyInput } from './CurrencyInput';
-export { default as DatePicker } from './DateTimePicker/DatePicker';
-export { default as MultiCheck } from './MultiCheck';
-export { default as Number } from './Number';
-export { default as Radio } from './Radio';
-export { default as Select } from './Select';
-export { default as Switch } from './Switch';
-export { default as Text } from './Text';
-export { default as Hidden } from './Hidden';
-export { default as TextArea } from './TextArea';
-export { default as TimePicker } from './DateTimePicker/TimePicker';
 
 export * from './EspressoButton';
 export * from './InlineEditInput';
